requirements.txt

fastapi==0.79.0
orjson
pydantic==1.9.0
uvicorn==0.18.2
# uvloop
tortoise-orm==0.17.3
asyncpg==0.26.0 

